requirements.txt
django==1.10.6
psycopg2==2.6.2
gunicorn==19.6.0
wagtail==1.9
django-ses==0.8.0
django-storages==1.5.2
boto3==1.4.4
facebook-sdk==2.0.0
django-allauth==0.29.0
elasticsearch==2.4.0
django-timezone-field==2.0
huey==1.3.0
redis==2.10.5
django-model-utils==3.0.0
django-simple-history==1.8.2
pytest==3.0.7
pytest-django==3.1.2
sympy==1.0
django-paypal==0.3.6
mixpanel==4.3.2
whitenoise==3.3.0
tweepy==3.5.0
django-fullurl==0.4
facebook-sdk==2.0.0
wagtailtinymce==4.2.1.5
django-autocomplete-light